Refactor mbtiles lib
* Broke up martin-mbtiles into multiple files
* Made all mbtiles functions take a `SqliteExecutor` -- this way they can be used with any SQLX connection structs - either a pool connection or an individual non-pooled connection.
* Simplified mbtiles bin a bit - I realized there is really no need to pretty print the output for the single value retrieval. Easier to just dump it to console as is.
* Bump martin-mbtiles to v0.2.0
* Minor fixes in tools docs, cargo.toml, and justfile
