+++ martin-mbtiles/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "martin-mbtiles"
-version = "0.1.0"
+version = "0.2.0"
 authors = ["Yuri Astrakhan <YuriAstrakhan@gmail.com>", "MapLibre contributors"]
 description = "A simple low-level MbTiles access and processing library, with some tile format detection and other relevant heuristics."
 keywords = ["mbtiles", "maps", "tiles", "mvt", "tilejson"]
@@ -16,13 +16,14 @@ license.workspace = true
 [dependencies]
 futures.workspace = true
 log.workspace = true
-martin-tile-utils = { path = "../martin-tile-utils", version = "0.1.0" }
+martin-tile-utils.workspace = true
 serde_json.workspace = true
 sqlx.workspace = true
 thiserror.workspace = true
 tilejson.workspace = true
 
 # Bin dependencies
+# TODO: Make these only enabled for the bin target, i.e. by using a feature, and add a just task to run the bin
 anyhow.workspace = true
 clap.workspace = true
 tokio.workspace = true
